Dropdown menu, search bar, popup description, message notification
Add to the HTML and CSS for the dropdown menu, add the code for the drop-down, and click off. Add some ID tags and made a shot list of search objects and add an alert message. Add to the HTML and CSS for the popup description bar. Add a subscribed message alert.